Add pxeless to docker image
Add pxeless aka create_image.sh to docker image as entrypoint. This
will make the container image on Docker Hub easier to use.

GNUPGHOME is set to TMPDIR, which should be safe as it is created
with u+rwx, to allow for non-root user to use the container.

Updated the README example accordingly.

By coincidence this also fixes #10